17-ProofofHuman/BioCheckTest_Backend/requirements.txt
Flask==2.3.3
cryptography==41.0.4
web3==6.11.0
python-dotenv==1.0.0
requests==2.31.0
py-solc-x==2.0.4
